- gcc 4 shutup fixes
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@998 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/network.h
@@ -99,7 +99,7 @@ class NetGame {
 	Game*		game;
 	
 	int		playernum;
-	uint		net_game_time;
+	ulong    net_game_time;
 	
 	uint		common_rand_seed;
